+++ common/clickhouse-srv/src/types/from_sql.rs
@@ -339,43 +339,3 @@ from_sql_impl! {
     f32: Float32,
     f64: Float64
 }
-
-#[cfg(test)]
-mod test {
-    use chrono::prelude::*;
-    use chrono_tz::Tz;
-
-    use crate::types::column::Either;
-    use crate::types::from_sql::FromSql;
-    use crate::types::DateTimeType;
-    use crate::types::SqlType;
-    use crate::types::ValueRef;
-
-    #[test]
-    fn test_u8() {
-        let v = ValueRef::from(42_u8);
-        let actual = u8::from_sql(v).unwrap();
-        assert_eq!(actual, 42_u8);
-    }
-
-    #[test]
-    fn test_bad_convert() {
-        let v = ValueRef::from(42_u16);
-        match u32::from_sql(v) {
-            Ok(_) => panic!("should fail"),
-            Err(e) => assert_eq!(
-                "From SQL error: `SqlType::UInt16 cannot be cast to u32.`".to_string(),
-                format!("{}", e)
-            ),
-        }
-    }
-
-    #[test]
-    fn null_to_datetime() {
-        let null_value = ValueRef::Nullable(Either::Left(
-            SqlType::DateTime(DateTimeType::DateTime32).into(),
-        ));
-        let date = Option::<DateTime<Tz>>::from_sql(null_value);
-        assert_eq!(date.unwrap(), None);
-    }
-}
