+++ common/clickhouse-srv/src/types/mod.rs
@@ -24,6 +24,7 @@ use chrono_tz::Tz;
 use hostname::get;
 use lazy_static::lazy_static;
 
+pub use self::block::decompress_buffer;
 pub use self::block::Block;
 pub use self::block::RCons;
 pub use self::block::RNil;
@@ -33,15 +34,17 @@ pub use self::block::Rows;
 pub use self::column::Column;
 pub use self::column::ColumnType;
 pub use self::column::Complex;
+pub use self::column::Either;
 pub use self::column::Simple;
-pub(crate) use self::date_converter::DateConverter;
+pub use self::column::StringPool;
+pub use self::date_converter::DateConverter;
+pub use self::decimal::decimal2str;
 pub use self::decimal::Decimal;
 pub use self::enums::Enum16;
 pub use self::enums::Enum8;
 pub use self::from_sql::FromSql;
 pub use self::from_sql::FromSqlResult;
-pub use self::options::Options;
-pub(crate) use self::options::OptionsSource;
+pub use self::options::*;
 pub use self::query::Query;
 pub(crate) use self::stat_buffer::StatBuffer;
 pub use self::value::Value;
