+++ common/clickhouse-srv/tests/it/types/block/compressed.rs
@@ -1,0 +1,34 @@
+// Copyright 2020 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_clickhouse_srv::types::decompress_buffer;
+
+#[test]
+fn test_decompress() {
+    let expected = vec![
+        1u8, 0, 2, 255, 255, 255, 255, 0, 1, 1, 1, 115, 6, 83, 116, 114, 105, 110, 103, 3, 97, 98,
+        99,
+    ];
+
+    let source = vec![
+        245_u8, 5, 222, 235, 225, 158, 59, 108, 225, 31, 65, 215, 66, 66, 36, 92, 130, 34, 0, 0, 0,
+        23, 0, 0, 0, 240, 8, 1, 0, 2, 255, 255, 255, 255, 0, 1, 1, 1, 115, 6, 83, 116, 114, 105,
+        110, 103, 3, 97, 98, 99,
+    ];
+
+    let mut cursor = std::io::Cursor::new(&source[..]);
+    let actual = decompress_buffer(&mut cursor, Vec::new()).unwrap();
+
+    assert_eq!(actual, expected);
+}
