+++ common/clickhouse-srv/tests/it/types/mod.rs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use self::encoder::Encoder;
-pub use self::parser::Parser;
-pub use self::read_ex::ReadEx;
-pub use self::uvarint::put_uvarint;
-
-mod encoder;
-mod parser;
-mod read_ex;
-mod uvarint;
+mod block;
+mod decimal;
+mod from_sql;
+mod opions;
+mod value;
+mod value_ref;
